Backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
chromadb
bs4
python-dotenv
